+++ src/Symfony/Component/Validator/GraphWalker.php
@@ -11,10 +11,8 @@
 
 namespace Symfony\Component\Validator;
 
-use Symfony\Component\Validator\ConstraintValidatorFactoryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
-use Symfony\Component\Validator\Mapping\ClassMetadataFactoryInterface;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Mapping\MemberMetadata;
 
@@ -24,29 +22,52 @@ use Symfony\Component\Validator\Mapping\MemberMetadata;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  * @author Bernhard Schussek <bschussek@gmail.com>
+ *
+ * @deprecated Deprecated since version 2.2, to be removed in 2.3. This class
+ *             has been replaced by {@link ValidationVisitorInterface} and
+ *             {@link MetadataInterface}.
  */
 class GraphWalker
 {
-    private $globalContext;
-    private $validatorFactory;
-    private $metadataFactory;
-    private $validatorInitializers = array();
-    private $validatedObjects = array();
+    /**
+     * @var ValidationVisitor
+     */
+    private $visitor;
 
-    public function __construct($root, ClassMetadataFactoryInterface $metadataFactory, ConstraintValidatorFactoryInterface $factory, array $validatorInitializers = array())
+    /**
+     * @var MetadataFactoryInterface
+     */
+    private $metadataFactory;
+
+    /**
+     * @var array
+     */
+    private $validatedObjects;
+
+    /**
+     * Creates a new graph walker.
+     *
+     * @param ValidationVisitor        $visitor
+     * @param MetadataFactoryInterface $metadataFactory
+     * @param array                    $validatedObjects
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
+     */
+    public function __construct(ValidationVisitor $visitor, MetadataFactoryInterface $metadataFactory, array &$validatedObjects = array())
     {
-        $this->globalContext = new GlobalExecutionContext($root, $this, $metadataFactory);
-        $this->validatorFactory = $factory;
+        $this->visitor = $visitor;
         $this->metadataFactory = $metadataFactory;
-        $this->validatorInitializers = $validatorInitializers;
+        $this->validatedObjects = &$validatedObjects;
     }
 
     /**
      * @return ConstraintViolationList
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
      */
     public function getViolations()
     {
-        return $this->globalContext->getViolations();
+        return $this->visitor->getViolations();
     }
 
     /**
@@ -57,128 +78,128 @@ class GraphWalker
      * @param object        $object       The object to validate
      * @param string        $group        The validator group to use for validation
      * @param string        $propertyPath
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
      */
     public function walkObject(ClassMetadata $metadata, $object, $group, $propertyPath)
-    {
-        foreach ($this->validatorInitializers as $initializer) {
-            if (!$initializer instanceof ObjectInitializerInterface) {
-                throw new \LogicException('Validator initializers must implement ObjectInitializerInterface.');
-            }
-            $initializer->initialize($object);
-        }
-
-        if ($group === Constraint::DEFAULT_GROUP && ($metadata->hasGroupSequence() || $metadata->isGroupSequenceProvider())) {
-            if ($metadata->hasGroupSequence()) {
-                $groups = $metadata->getGroupSequence();
-            } else {
-                $groups = $object->getGroupSequence();
-            }
-
-            foreach ($groups as $group) {
-                $this->walkObjectForGroup($metadata, $object, $group, $propertyPath, Constraint::DEFAULT_GROUP);
-
-                if (count($this->getViolations()) > 0) {
-                    break;
-                }
-            }
-        } else {
-            $this->walkObjectForGroup($metadata, $object, $group, $propertyPath);
-        }
-    }
-
-    protected function walkObjectForGroup(ClassMetadata $metadata, $object, $group, $propertyPath, $propagatedGroup = null)
     {
         $hash = spl_object_hash($object);
 
         // Exit, if the object is already validated for the current group
         if (isset($this->validatedObjects[$hash][$group])) {
-                return;
+            return;
         }
 
         // Remember validating this object before starting and possibly
         // traversing the object graph
         $this->validatedObjects[$hash][$group] = true;
 
-        $currentClass = $metadata->getClassName();
-
-        foreach ($metadata->findConstraints($group) as $constraint) {
-            $this->walkConstraint($constraint, $object, $group, $propertyPath, $currentClass);
-        }
-
-        if (null !== $object) {
-            $pathPrefix = empty($propertyPath) ? '' : $propertyPath.'.';
-            foreach ($metadata->getConstrainedProperties() as $property) {
-                $this->walkProperty($metadata, $property, $object, $group, $pathPrefix.$property, $propagatedGroup);
-            }
-        }
+        $metadata->accept($this->visitor, $object, $group, $propertyPath);
     }
 
+    protected function walkObjectForGroup(ClassMetadata $metadata, $object, $group, $propertyPath, $propagatedGroup = null)
+    {
+        $metadata->accept($this->visitor, $object, $group, $propertyPath, $propagatedGroup);
+    }
+
+    /**
+     * Validates a property of a class.
+     *
+     * @param Mapping\ClassMetadata $metadata
+     * @param                       $property
+     * @param                       $object
+     * @param                       $group
+     * @param                       $propertyPath
+     * @param null                  $propagatedGroup
+     *
+     * @throws Exception\UnexpectedTypeException
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
+     */
     public function walkProperty(ClassMetadata $metadata, $property, $object, $group, $propertyPath, $propagatedGroup = null)
     {
+        if (!is_object($object)) {
+            throw new UnexpectedTypeException($object, 'object');
+        }
+
         foreach ($metadata->getMemberMetadatas($property) as $member) {
-            $this->walkMember($member, $member->getValue($object), $group, $propertyPath, $propagatedGroup);
+            $member->accept($this->visitor, $member->getValue($object), $group, $propertyPath, $propagatedGroup);
         }
     }
 
+    /**
+     * Validates a property of a class against a potential value.
+     *
+     * @param Mapping\ClassMetadata $metadata
+     * @param                       $property
+     * @param                       $value
+     * @param                       $group
+     * @param                       $propertyPath
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
+     */
     public function walkPropertyValue(ClassMetadata $metadata, $property, $value, $group, $propertyPath)
     {
         foreach ($metadata->getMemberMetadatas($property) as $member) {
-            $this->walkMember($member, $value, $group, $propertyPath);
+            $member->accept($this->visitor, $value, $group, $propertyPath);
         }
     }
 
     protected function walkMember(MemberMetadata $metadata, $value, $group, $propertyPath, $propagatedGroup = null)
     {
-        $currentClass = $metadata->getClassName();
-        $currentProperty = $metadata->getPropertyName();
-
-        foreach ($metadata->findConstraints($group) as $constraint) {
-            $this->walkConstraint($constraint, $value, $group, $propertyPath, $currentClass, $currentProperty);
-        }
-
-        if ($metadata->isCascaded()) {
-            $this->walkReference($value, $propagatedGroup ?: $group, $propertyPath, $metadata->isCollectionCascaded(), $metadata->isCollectionCascadedDeeply());
-        }
+        $metadata->accept($this->visitor, $value, $group, $propertyPath, $propagatedGroup);
     }
 
+    /**
+     * Validates an object or an array.
+     *
+     * @param      $value
+     * @param      $group
+     * @param      $propertyPath
+     * @param      $traverse
+     * @param bool $deep
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
+     */
     public function walkReference($value, $group, $propertyPath, $traverse, $deep = false)
     {
-        if (null !== $value) {
-            if (!is_object($value) && !is_array($value)) {
-                throw new UnexpectedTypeException($value, 'object or array');
-            }
-
-            if ($traverse && (is_array($value) || $value instanceof \Traversable)) {
-                foreach ($value as $key => $element) {
-                    // Ignore any scalar values in the collection
-                    if (is_object($element) || is_array($element)) {
-                        // Only repeat the traversal if $deep is set
-                        $this->walkReference($element, $group, $propertyPath.'['.$key.']', $deep, $deep);
-                    }
-                }
-            }
-
-            if (is_object($value)) {
-                $metadata = $this->metadataFactory->getClassMetadata(get_class($value));
-                $this->walkObject($metadata, $value, $group, $propertyPath);
-            }
-        }
+        $this->visitor->validate($value, $group, $propertyPath, $traverse, $deep);
     }
 
+    /**
+     * Validates a value against a constraint.
+     *
+     * @param Constraint $constraint
+     * @param            $value
+     * @param            $group
+     * @param            $propertyPath
+     * @param null       $currentClass
+     * @param null       $currentProperty
+     *
+     * @deprecated Deprecated since version 2.2, to be removed in 2.3.
+     */
     public function walkConstraint(Constraint $constraint, $value, $group, $propertyPath, $currentClass = null, $currentProperty = null)
     {
-        $validator = $this->validatorFactory->getInstance($constraint);
+        $metadata = null;
 
-        $localContext = new ExecutionContext(
-            $this->globalContext,
+        // BC code to make getCurrentClass() and getCurrentProperty() work when
+        // called from within this method
+        if (null !== $currentClass) {
+            $metadata = $this->metadataFactory->getMetadataFor($currentClass);
+
+            if (null !== $currentProperty && $metadata instanceof PropertyMetadataContainerInterface) {
+                $metadata = current($metadata->getPropertyMetadata($currentProperty));
+            }
+        }
+
+        $context = new ExecutionContext(
+            $this->visitor,
+            $metadata,
             $value,
-            $propertyPath,
             $group,
-            $currentClass,
-            $currentProperty
+            $propertyPath
         );
 
-        $validator->initialize($localContext);
-        $validator->validate($value, $constraint);
+        $context->validateValue($value, $constraint);
     }
 }